fix(frontend): Handle empty error message in toast + hide GridEditorMenu when the component is not visible
