+++ frontend/src/lib/components/apps/components/helpers/RunnableWrapper.svelte
@@ -186,12 +186,10 @@
 				let message = sideEffect?.configuration?.sendErrorToast?.message
 				const appendError = sideEffect?.configuration?.sendErrorToast?.appendError
 
+				if (!message) return
+
 				if (typeof message === 'function') {
 					message = await message()
-				}
-
-				if (!message) {
-					message = 'An error occurred.'
 				}
 
 				sendUserToast(message, true, [], appendError ? errorMessage : undefined)
